refactor: separate out decoding and executing instructions
